fix: enabled the rule import/no-cycle
